app/requirements.txt
anyio==3.6.2
black==22.6.0
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.0
fastapi==0.87.0
flake8==4.0.1
h11==0.14.0
idna==3.4
importlib-metadata==5.0.0
influxdb-client==1.35.0
jwt==1.3.1
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.21.6
opencv-python==4.6.0.66
pathspec==0.11.0
platformdirs==2.6.2
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.10.2
pyflakes==2.4.0
python-dateutil==2.8.2
python-dotenv==0.21.0
reactivex==4.0.4
six==1.16.0
sniffio==1.3.0
starlette==0.21.0
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
zipp==3.10.0
